Allows user to wait for security check
By using port fowarding and VNC, we can interact with the selenium browser
in real time on our local machine. Details are defined in the readme.

Additionally, an email is sent out when a security check times out. This will
alert us to log on and manually fix the security check.

Since chrome profile data is mounted as a volume to the container,
we should not have to worry about these security checks often as the
chrome profile is used to log in